fix(follow-me): Moves follow-me state to its feature.
