+++ react/features/follow-me/subscriber.ts
@@ -17,7 +17,7 @@ import { FOLLOW_ME_COMMAND } from './constants';
  * notify all listeners.
  */
 StateListenerRegistry.register(
-    /* selector */ state => state['features/base/conference'].followMeEnabled,
+    /* selector */ state => state['features/follow-me'].followMeEnabled,
     /* listener */ (newSelectedValue, store) => _sendFollowMeCommand(newSelectedValue || 'off', store));
 
 /**
@@ -88,7 +88,7 @@ function _getFollowMeState(state: IReduxState) {
     const stageFilmstrip = isStageFilmstripEnabled(state);
 
     return {
-        recorder: state['features/base/conference'].followMeRecorderEnabled,
+        recorder: state['features/follow-me'].followMeRecorderEnabled,
         filmstripVisible: state['features/filmstrip'].visible,
         maxStageParticipants: stageFilmstrip ? state['features/base/settings'].maxStageParticipants : undefined,
         nextOnStage: pinnedParticipant?.id,
@@ -130,7 +130,7 @@ function _sendFollowMeCommand(
         );
 
         return;
-    } else if (!state['features/base/conference'].followMeEnabled) {
+    } else if (!state['features/follow-me'].followMeEnabled) {
         return;
     }
